Add login decorators and admin function security

+++ products/views.py
@@ -1,5 +1,6 @@
 from django.shortcuts import render, get_object_or_404, reverse, redirect
 from django.contrib import messages
+from django.contrib.auth.decorators import login_required
 from django.db.models import Q
 from django.db.models.functions import Lower
 from .models import Product, Category
@@ -9,7 +10,9 @@ from .forms import ProductForm
 
 
 def all_products(request):
-    """ A view to show all products, including sorting and searching """
+    """ 
+    A view to show all products, including sorting and searching
+    """
 
     products = Product.objects.all()
     query = None
@@ -60,7 +63,9 @@ def all_products(request):
 
 
 def product_detail(request, product_id):
-    """ A view to show details of an individual product """
+    """
+    Show details of an individual product
+    """
 
     product = get_object_or_404(Product, pk=product_id)
 
@@ -69,10 +74,16 @@ def product_detail(request, product_id):
     }
     return render(request, 'products/product_detail.html', context)
 
+
+@login_required
 def add_product(request):
     """
     Add a product to the store (admin only)
     """
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admins can do that.')
+        return redirect(reverse('home'))
+
     if request.method == 'POST':
         form = ProductForm(request.POST, request.FILES)
         if form.is_valid():
@@ -93,10 +104,16 @@ def add_product(request):
     return render(request, template, context)
 
 
+@login_required
 def edit_product(request, product_id):
     """
     Edit a product in the store (admin only)
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admins can do that.')
+        return redirect(reverse('home')
+
     product = get_object_or_404(Product, pk=product_id)
     form = ProductForm(instance=product)
 
@@ -121,10 +138,16 @@ def edit_product(request, product_id):
     return render(request, template, context)
 
 
+@login_required
 def delete_product(request, product_id):
     """ 
     Delete a product from the store (admin only)
     """
+
+    if not request.user.is_superuser:
+        messages.error(request, 'Sorry, only store admins can do that.')
+        return redirect(reverse('home')
+
     product = get_object_or_404(Product, pk=product_id)
     product.delete()
     messages.success(request, 'Product deleted!')
